Move list-subscribers endpoint to folder
